requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
boto3==1.9.188
botocore==1.12.188
chardet==3.0.4
Click==7.0
docutils==0.14
idna==2.8
jmespath==0.9.4
multidict==4.5.2
python-dateutil==2.8.0
python-lambda==3.2.6
PyYAML==5.1.1
s3transfer==0.2.1
six==1.12.0
slackclient==2.1.0
urllib3==1.25.3
yarl==1.3.0
